Optionally add attachments to the end of the message
Added a is_attachment arg to self.attach().
If False (the default) the attachment is added immediately to the message (current behaviour preserved).
If True, the attachment is added to self.__attachments_data instead.

In self.prepare() if self.__attachments_data in not empty, add attachments to message (using self.attch with is_attachment=False).

Relevant to #4